Add a link to add FreshRSS as a Firefox feed reader
This is more a proof of concept (it works but has some flaws).

The major flaw is the URL generated. It needs to be an absolute URL so it is not possible to use the helper to generate it.
The URL generated by the static method introduce an extra / which breaks the UI. It is possible to add the feed but we will
loose the CSS. To correct than we need to fix the URL generation in Minz.
I found three places where the fix can be implemented:
  - Minz_Url:64 the separator must be ? instead of /?
  - Minz_Request:103 the return URL shouldn't end with /
  - Minz_Request:105 the return URL shouldn't end with /

The other flaws are the location of the link in the page and translation.