Change behavior when selecting an article
See #524
Before, when the configuration was set so the article is marked as read when selected, if you mark it as unread while it is open and then close it, it is marked again as read. This behavior was not straightforward.
Now, the article is not marked again as read.