Fix read all shortcut
Before, the read all shortcut raised a javascript exception when we use the confirmation option. The confirmation was not even displayed. The code was failling since there was a missing variable.
Now, the shortcut ask for confirmation when the option is selected in the reading configuration page.

The definition of the shortcut was buggy since the confirmation was handled properly in an other location. I simplified the code by removing the confirmation code in the shortcut.