Change statistic average
I was using the stats and it feels that the stat average was useless in that form since the line is always at the same position no matter what is the value.
So I deleted that line and added the average in the title. It is way more useful this way.

I updated some translations both on i18n files and on http://i18n.freshrss.org