Add a way to parse search string to extract keywords
This feature is not in use at the moment, but it will be handy to reorganize the query building process. It allows to have more than one keyword in the search box.
Full tests are available as well.

It probably needs a refactoring later, but I think this is the first step to make the application full object oriented and testable.