Harmonize share configuration view.
Before, for shares that don't need options, only a button to remove it was visible. It was source of confusion for users.
I changed the look of those shares by using the same layout as others (minus the help). As there is no configuration possible for the url, the field is disabled but it is possible to change the name of the share.

See #787