Fix entry DAO query usage
I did not fix the call in the previous commit. I hope this one is the last change needed.

We definitely need a templating engine so we could use the same controller to output different things.
This will remove code duplication between the api and the web interface.
It will allows us to build other type of api, and also refactor the rss feed as a different view of the same dataset.