Fix feed and category side effect
Before, when deleting a feed or a category, the user queries were deleted as well. No matter if they were related or not.
Now, they are deleted only if they are related.

I this this fix is not the best way to handle that. I think it would be better if we could find a way to create a UserQuery object from the array.
The same applies when displaying the user queries in the interface.

See #980