Add support for a new shortcut in reader view
Now you can open the original page in the reader view with the same shortcut you'll use in the normal view.
I've changed how we identify the link to make it more flexible.
The previous way was too restrictive since the selector used a really strict path to get the url.

There was another way to achieve the same thing without changing the selector.
It was quite ugly since some meaningless class would be added on the markup to match the selector query.

See #1400