Docker image alternative using Ubuntu
Ubuntu seems to be faster, cf. e.g. PHPBench on:

https://www.phoronix.com/scan.php?page=article&item=freebsd-12-windows&num=5

https://www.phoronix.com/scan.php?page=article&item=docker-summer-2018&num=4
This image is also different from our Alpine-based image as it is using
Apache Event and PHP-FPM as FastCGI. They is also supposed to be faster
for some types of workload than mod_php but with other drawbacks (poorer
control of flush, more processes running...).
So some concreate performance tests would be needed to check whether
this is relevant at all.
It is quite usual for Docker images to offer both a Debian/Ubuntu and an
Alpine version, so we could also do the same if there is any need.
Follow-up of https://github.com/FreshRSS/FreshRSS/pull/2202