Greener Travis?
With PHP 7.4 just around the corner, I suggest to test only the minimum
and maximum PHP versions we support in Travis, to save resources
(greener).
For the translation check, with can test with max version -1 to run on
yet another PHP version.
It makes also the Travis tests faster to complete.