volume names include redundant prefix (#3108)
This proposed change includes tweaks to the names of the services and volumes, and adds an explicit label to the postgres and freshrss containers.

Using a more generic "freshrss-db" instead of "freshrss_postgresql" seems more standard among other docker projects and makes it a bit easier to switch databases later.

Removing the "freshrss_" prefix from the volume names solves a problem where the docker-compose automatically prepends a "project name" to volume names upon running "up".  So if your docker-compose.yml file is stored in a folder named "freshrss", you would end up with a redundant volume name of "freshrss_freshrss_data".

This also adds a restart policy to the db container.