Change base extension class type
Before, there was some guidelines on how to use the extension class
and how to extend it. Those guidelines were defined as comments.
Now, those guidelines are enforced by the code itself. There is no
need for those comments anymore.

Beware, this change will break every extension that are not following
those guideline. I've made the change for the core extensions but works
need to be done for the others.